backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
chromadb==0.4.22
llama-index==0.10.68
llama-index-vector-stores-chroma==0.1.4
llama-index-embeddings-huggingface==0.1.4
datasets==2.16.1
huggingface-hub==0.20.3
pydantic==2.9.2
python-multipart==0.0.6
openai==1.10.0
